refactor: rebase and rename class router
