+++ composer.json
@@ -4,7 +4,8 @@
     "psr-4": {
       "controllers\\": "controllers",
       "models\\": "models",
-      "views\\": "views"
+      "views\\": "views",
+      "app\\": "app"
     }
   }
 }
