+++ public/index.php
@@ -3,10 +3,10 @@
 declare(strict_types=1);
 require_once dirname(__DIR__) . '/vendor/autoload.php';
 
-use controllers\RoutingController;
+use app\Router;
 
-$routingController = new RoutingController();
-$routingController->route($_SERVER['REQUEST_URI']);
+$router = new Router();
+$router->route($_SERVER['REQUEST_URI']);
 
 // echo "Вы находитесь по адресу: ".parse_ugit branchrl($url, PHP_URL_PATH);
