+++ views/IndexView.php
@@ -5,6 +5,7 @@ namespace views;
 class IndexView
 {
     private array $posts;
+
     public function __construct(array $posts)
     {
         $this->posts = $posts;
